feat: split Badge and EntityStatusBadge
