+++ packages/components/badge/src/index.ts
@@ -1,2 +1,5 @@
 export { Badge } from './Badge';
-export type { BadgeProps, BadgeVariant } from './types';
+export type { BadgeProps } from './Badge';
+export { EntityStatusBadge } from './EntityStatusBadge';
+export type { EntityStatusBadgeProps } from './EntityStatusBadge';
+export type { BadgeVariant } from './types';
